Se quito app.js y se comenzo a usar index.js, tambien se siguio con la estructura html y el css
